requirements.txt
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.28
certifi==2021.10.8
charset-normalizer==2.0.12
future==0.18.2
idna==3.3
install==1.3.5
iso8601==1.0.2
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.0
opencv-python==4.5.5.64
pyserial==3.5
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.8
